add snapshot_iterator_test.go
